+++ src/utils/mod.rs
@@ -1,0 +1,9 @@
+//! A collection of audio utilities.
+
+pub(crate) mod entity_set;
+#[cfg(any(feature = "profiling", test))]
+pub(crate) mod profiling;
+
+pub mod fixed_vec;
+pub mod perceptual_volume;
+pub mod timeline;
